TestingPower -> BrowserEfficiencyTest
Changed the solution, project, and namespace names to
BrowserEfficiencyTest to facilitate scope increase beyond power
