+++ BrowserEfficiencyTest/E3Events.cs
@@ -1,6 +1,6 @@
 ﻿//--------------------------------------------------------------
 //
-// Microsoft Edge Power Test
+// Browser Efficiency Test
 // Copyright(c) Microsoft Corporation
 // All rights reserved.
 //
@@ -27,7 +27,7 @@
 
 using System;
 
-namespace TestingPower
+namespace BrowserEfficiencyTest
 {
     /// <summary>
     /// Data container representing the event Microsoft-Windows-Energy-Estimation-Engine/QueryStats/EnergyEstimate.
